+++ packages/core/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log - @sap/guided-answers-extension-core
 
+## 0.10.0
+
+### Minor Changes
+
+-   ed0fd7c: Add filters menu for tutorials
+
+### Patch Changes
+
+-   Updated dependencies [ed0fd7c]
+    -   @sap/knowledge-hub-extension-types@0.10.0
+
 ## 0.9.3
 
 ### Patch Changes